feat(portal-accounts): add request and real-time limit columns and freeze action column
