+++ packages/i18n/src/locales/zh-CN.json
@@ -3605,6 +3605,7 @@
     "reset_history_actor_unknown": "未知",
     "account_permission_profile": "账户权限模板",
     "account_permission_profile_placeholder": "选择账户权限模板",
+    "concurrency_limit_col": "并发",
     "quota_on_account_hint": "限额与模型/渠道权限挂在账号上，由该账号下全部 Key 共享。",
     "last_login": "最近登录",
     "create": "创建用户",
