+++ scripts/file-size-baseline.json
@@ -21,7 +21,7 @@
     "pages/auth-files/hooks/useAuthFilesDetailEditors.ts": 1318,
     "pages/auth-files/hooks/useAuthFilesFilesPresentation.tsx": 967,
     "pages/auth-files/hooks/useAuthFilesStatusState.ts": 1298,
-    "pages/end-users/EndUsersPage.tsx": 1126,
+    "pages/end-users/EndUsersPage.tsx": 908,
     "pages/identity-fingerprint/IdentityFingerprintPage.tsx": 1832,
     "pages/image-generation/components/ImageGenerationPageContent.tsx": 1191,
     "pages/models/ModelsPage.tsx": 1251,
